need to strip trailing "space" character
even if it's not whitespace. so strip() with default arg (whitespace) won't work for all cases.

FIXME: need to reuse whatever space character (in a class or list of word-separating characters) was used in the original, provided sentence rather than replacing it with a space always. e.g. don't replace linefeeds with spaces!